+++ tests/expectations/core/test_expect_column_value_z_scores_to_be_less_than.py
@@ -51,11 +51,10 @@ def test_sa_expect_column_value_z_scores_to_be_less_than_impl(postgresql_engine)
         },
     )
     expectation = ExpectColumnValueZScoresToBeLessThan(expectationConfiguration)
-    batch_data = SqlAlchemyBatchData(
-        engine=postgresql_engine, table_name="z_score_test_data"
-    )
-    engine = SqlAlchemyExecutionEngine(
-        engine=postgresql_engine, batch_data_dict={"my_id": batch_data}
+    engine = SqlAlchemyExecutionEngine(engine=postgresql_engine)
+    engine.load_batch_data(
+        "my_id",
+        SqlAlchemyBatchData(execution_engine=engine, table_name="z_score_test_data"),
     )
     result = expectation.validate(Validator(execution_engine=engine))
     assert result == ExpectationValidationResult(